rename gr2r for r2r for backwards compatibility
