+++ deepinv/loss/r2r.py
@@ -1,36 +1,48 @@
 from __future__ import annotations
 from typing import Union
 import torch
+import math
 from deepinv.loss.loss import Loss
 from deepinv.loss.metric.metric import Metric
+from deepinv.physics.noise import GaussianNoise, PoissonNoise, GammaNoise
+from deepinv.physics.noise import NoiseModel
 
 
 class R2RLoss(Loss):
     r"""
-    Recorrupted-to-Recorrupted (R2R) Loss
+    Generalized Recorrupted-to-Recorrupted (GR2R) Loss
 
-    This loss can be used for unsupervised image denoising with unorganized noisy images.
-
-    The loss is designed for the noise model:
-
-    .. math::
-
-        y \sim\mathcal{N}(u,\sigma^2 I) \quad \text{with}\quad u= A(x).
-
-    The loss is computed as:
+    This loss can be used for unsupervised image denoising with unorganized noisy images, where
+    the noise model :math:`y\sim p(y\vert x)` belongs to the exponential family as
 
     .. math::
 
-        \| y^- - AR(y^+) \|_2^2 \quad \text{s.t.} \quad y^+ = y + \alpha z, \quad y^- = y -  z / \alpha
+        p(y\vert x) =  h(x) \exp \left( y^\top \eta(x) - \phi(x) \right)
+
+    For this family of noisy measurements, we genealize the corruption strategy as
+
+    .. math::
+
+        y_1 \sim p(y_1 \vert y, \alpha ),
+
+    .. math::
+
+        y_2 = \frac{1}{\alpha} \left( y - y_1(1-\alpha) \right),
+
+    then, the loss is computed as:
+
+    .. math::
+
+        \| AR(y_1) - y_2 \|_2^2
 
     where :math:`R` is the trainable network, :math:`A` is the forward operator,
-    :math:`y` is the noisy measurement,
-    :math:`z` is the additional Gaussian noise of standard deviation :math:`\sigma`,
-    and :math:`\alpha` is a scaling factor.
+    :math:`y` is the noisy measurement, and :math:`\alpha` is a scaling factor.
 
 
-    This loss is statistically equivalent to the supervised loss function defined on noisy/clean image pairs
-    according to authors in https://ieeexplore.ieee.org/document/9577798
+    This loss generalizes the Recorrupted-to-Recorrupted (R2R) loss for the exponential family of noise models
+    as presented in https://arxiv.org/abs/2412.04648. For the gaussian noise model, the R2R loss is recovered
+    and is statistically equivalent to the supervised loss function defined on noisy/clean image pairs
+    according to authors in https://ieeexplore.ieee.org/document/9577798.
 
     .. warning::
 
@@ -38,80 +50,72 @@ class R2RLoss(Loss):
 
     .. note::
 
-        :math:`\sigma` should be chosen equal or close to :math:`\sigma` to obtain the best performance.
-
-    .. note::
-
         To obtain the best test performance, the trained model should be averaged at test time
-        over multiple realizations of the added noise, i.e. :math:`\hat{x} = \frac{1}{N}\sum_{i=1}^N R(y+\alpha z_i)`
+        over multiple realizations of the added noise, i.e. :math:`\hat{x} = \frac{1}{N}\sum_{i=1}^N R(y_1^{(i)})`
         where :math:`N>1`. This can be achieved using :meth:`adapt_model`.
 
+    .. deprecated:: 0.2.3
+
+        The ``sigma`` paramater is deprecated and will be removed in future versions. Use ``noise_model`` parameter instead.
+
     :param Metric, torch.nn.Module metric: metric for calculating loss, defaults to MSE.
-    :param float sigma: standard deviation of the Gaussian noise used for the perturbation.
-    :param float alpha: scaling factor of the perturbation.
+    :param NoiseModel noise_model: noise model of the exponential family, defaults to Gaussian.
+    :param float alpha: scaling factor of the corruption.
     :param int eval_n_samples: number of samples used for the Monte Carlo approximation.
-
-    |sep|
-
-    :Example:
-
-        >>> import torch
-        >>> import deepinv as dinv
-        >>> sigma = 0.1
-        >>> physics = dinv.physics.Denoising(dinv.physics.GaussianNoise(sigma))
-        >>> model = dinv.models.MedianFilter()
-        >>> loss = dinv.loss.R2RLoss(sigma=sigma, eval_n_samples=2)
-        >>> model = loss.adapt_model(model) # important step!
-        >>> x = torch.ones((1, 1, 8, 8))
-        >>> y = physics(x)
-        >>> x_net = model(y, physics, update_parameters=True) # save extra noise in forward pass
-        >>> l = loss(x_net, y, physics, model)
-        >>> print(l.item() > 0)
-        True
 
     """
 
     def __init__(
         self,
         metric: Union[Metric, torch.nn.Module] = torch.nn.MSELoss(),
-        sigma=0.1,
+        noise_model: NoiseModel = GaussianNoise(0.1),
         alpha=0.5,
+        sigma=None,
         eval_n_samples=5,
     ):
         super(R2RLoss, self).__init__()
-        self.name = "r2r"
+        self.name = "gr2r"
         self.metric = metric
-        self.sigma = sigma
         self.alpha = alpha
         self.eval_n_samples = eval_n_samples
+        self.noise_model = noise_model
+
+        if sigma is not None:
+
+            print(
+                "The sigma parameter is deprecated and will be removed in future versions. "
+                "Please use the noise_model parameter instead."
+            )
+
+            self.noise_model = GaussianNoise(sigma)
 
     def forward(self, x_net, y, physics, model, **kwargs):
         r"""
-        Computes the R2R Loss.
+        Computes the GR2R Loss.
 
         :param torch.Tensor y: Measurements.
         :param deepinv.physics.Physics physics: Forward operator associated with the measurements.
         :param torch.nn.Module model: Reconstruction model.
-        :return: (torch.Tensor) R2R loss.
+        :return: (torch.Tensor) GR2R loss.
         """
 
-        pert = model.get_noise()
-        y_minus = y - pert / self.alpha
-        return self.metric(physics.A(x_net), y_minus)
+        y1 = model.get_corruption()
+        y2 = (1 / self.alpha) * (y - y1 * (1 - self.alpha))
+        return self.metric(physics.A(x_net), y2)
 
-    def adapt_model(self, model: torch.nn.Module) -> R2RModel:
+    def adapt_model(self, model, **kwargs):
         r"""
         Adds noise to model input.
 
 
         This method modifies a reconstruction
-        model :math:`R` to include the splitting mechanism at the input:
+        model :math:`R` to include the re-corruption mechanism at the input:
 
         .. math::
 
-            \hat{R}(y) = \frac{1}{N}\sum_{i=1}^N R(y+\alpha z_i)
+            \hat{R}(y) = \frac{1}{N}\sum_{i=1}^N R(y_1^{(i)})
 
-        where :math:`N\geq 1` are the number of samples used for the Monte Carlo approximation.
+        where :math:`y_1^{(i)} \sim p(y_1 \vert y, \alpha)` are i.i.d samples, and :math:`N\geq 1` are the number of samples used for the Monte Carlo approximation.
         During training (i.e. when ``model.train()``), we use only one sample, i.e. :math:`N=1`
         for computational efficiency, whereas at test time, we use multiple samples for better performance.
 
@@ -120,36 +124,88 @@ class R2RLoss(Loss):
         :param float alpha: scaling factor of the perturbation.
         :return: (torch.nn.Module) Modified model.
         """
-        if isinstance(model, R2RModel):
-            return model
-        else:
-            return R2RModel(model, self.sigma, self.alpha, self.eval_n_samples)
+
+        return R2RModel(
+            model, self.noise_model, self.alpha, self.eval_n_samples, **kwargs
+        )
+
+
+def set_gaussian_corruptor(y, alpha, sigma):
+    mu = torch.ones_like(y) * 0.0
+    sigma = torch.ones_like(y) * sigma
+    sampler = torch.distributions.Normal(mu, sigma)
+    corruptor = lambda: y + sampler.sample() * (math.sqrt(alpha / (1 - alpha)))
+    return corruptor
+
+
+def set_binomial_corruptor(y, alpha, gamma):
+    z = y / gamma
+    sampler = torch.distributions.Binomial(torch.round(z), alpha)
+    corruptor = lambda: gamma * (z - sampler.sample()) / (1 - alpha)
+    return corruptor
+
+
+def set_beta_corruptor(y, alpha, l):
+    tmp = torch.ones_like(y)
+    concentration1 = tmp * l * alpha
+    concentration0 = tmp * l * (1 - alpha)
+    sampler = torch.distributions.Beta(concentration1, concentration0)
+    corruptor = lambda: y * (1 - sampler.sample()) / (1 - alpha)
+    return corruptor
 
 
 class R2RModel(torch.nn.Module):
-    def __init__(self, model, sigma, alpha, eval_n_samples):
+    r"""
+    Generalized Recorrupted-to-Recorrupted (GR2R) Model
+
+    """
+
+    def __init__(self, model, noise_model, alpha, eval_n_samples):
         super(R2RModel, self).__init__()
         self.model = model
-        self.extra_noise = 0
-        self.sigma = sigma
-        self.alpha = alpha
+        self.noise_model = noise_model
         self.eval_n_samples = eval_n_samples
+        self.alpha = alpha
 
     def forward(self, y, physics, update_parameters=False):
-        eval_n_samples = 1 if self.training else self.eval_n_samples
 
+        eval_n_samples = 1 if self.training else self.eval_n_samples
         out = 0
+        corruptor = self.get_corruptor(y)
+
         with torch.set_grad_enabled(self.training):
             for i in range(eval_n_samples):
-                extra_noise = torch.randn_like(y) * self.sigma
-                y_plus = y + extra_noise * self.alpha
-                out += self.model(y_plus, physics)
+                y1 = corruptor()
+                out += self.model(y1, physics)
 
-            if self.training and update_parameters:  # save the noise
-                self.extra_noise = extra_noise
+            if self.training and update_parameters:
+                self.corruption = y1
 
             out = out / eval_n_samples
         return out
 
-    def get_noise(self):
-        return self.extra_noise
+    def get_corruptor(self, y):
+        alpha = self.alpha
+
+        if isinstance(self.noise_model, GaussianNoise):
+
+            sigma = self.noise_model.sigma
+            return set_gaussian_corruptor(y, alpha, sigma)
+
+        elif isinstance(self.noise_model, PoissonNoise):
+
+            gain = self.noise_model.gain
+            return set_binomial_corruptor(y, alpha, gain)
+
+        elif isinstance(self.noise_model, GammaNoise):
+
+            l = self.noise_model.l
+            return set_beta_corruptor(y, alpha, l)
+
+        else:
+            raise ValueError(
+                f"Noise model {self.noise_model} not supported, available options are Gaussian, Poisson and Gamma."
+            )
+
+    def get_corruption(self):
+        return self.corruption
